refactor: 拆分 modern_settings_dialog.py（4811→1857 行）
纯机械搬移，零行为变更（AST 逐字节校验）：
- 控件库 44 符号 → pet/settings_widgets.py（复活批6-7 孤儿文件名，
  这次真正接线；含 Masonry 渲染 helper 与设置弹窗控件）
- MenuLayoutEditor → pet/settings_menu_layout_editor.py
- _AiSettingsPage → pet/chat/ai_settings_page.py（回归 chat 域）
- 主题 QSS 常量与 _settings_stylesheet → pet/settings_theme_qss.py

modern_settings_dialog.py 保留 re-export，pet/ 调用方零改动；tests 仅
3 处 monkeypatch 目标随真实命名空间迁移（断言零改动）。
tests/test_architecture.py 新增 MODERN_SETTINGS_DIALOG_PY_LINE_BUDGET
=1907，堵上「只有 window.py 受行数约束」的制度漏洞。

验证：全量 1308 passed / 7 skipped（仅 1 本机既有失败）；ruff 干净；
聚焦 120 passed（本人复跑）。
