+++ tests/test_requested_regressions.py
@@ -955,11 +955,11 @@ def test_settings_stylesheet_has_dark_overrides(monkeypatch):
     """深色系统下新版设置必须追加深色覆盖段（白底白字不可读问题）。"""
     from pet.modern_settings_dialog import _settings_stylesheet
 
-    monkeypatch.setattr("pet.modern_settings_dialog._system_dark", lambda: True)
+    monkeypatch.setattr("pet.settings_theme_qss._system_dark", lambda: True)
     qss = _settings_stylesheet()
     assert "background: #202024" in qss
     assert "color: #e4e4e9" in qss
-    monkeypatch.setattr("pet.modern_settings_dialog._system_dark", lambda: False)
+    monkeypatch.setattr("pet.settings_theme_qss._system_dark", lambda: False)
     qss_light = _settings_stylesheet()
     assert "background: #202024" not in qss_light
     # 浅色也必须显式给按钮补文字色（防深色 palette 白字）
@@ -996,7 +996,7 @@ def test_settings_window_rethemes_immediately_with_the_appearance_selector(
     from pet.config import Config
 
     app = QApplication.instance() or QApplication([])
-    monkeypatch.setattr(settings_mod, "_system_dark", lambda: False)
+    monkeypatch.setattr("pet.settings_theme_qss._system_dark", lambda: False)
     monkeypatch.setattr(settings_mod.autostart_mod, "is_enabled", lambda: False)
     dialog = settings_mod.ModernSettingsDialog(Config(tmp_path), include_ai=False)
     assert "QDialog { background: #202024" not in dialog.styleSheet()
